+++ soupsieve/__init__.py
@@ -30,10 +30,10 @@ from .__meta__ import __version__, __version_info__  # noqa: F401
 from . import css_parser as cp
 from . import css_match as cm
 from . import css_types as ct
-from .util import DEBUG
+from .util import DEBUG, _QUIRKS  # noqa: F401
 
 __all__ = (
-    'SoupSieve', 'compile', 'purge', 'DEBUG',
+    'SoupSieve', 'compile', 'purge', 'DEBUG', "_QUIRKS"
     'comments', 'icomments', 'closest', 'select', 'select_one',
     'iselect', 'match', 'filter'
 )