+++ soupsieve/util.py
@@ -4,8 +4,11 @@ from functools import wraps
 import warnings
 import sys
 import struct
+import os
+MODULE = os.path.dirname(__file__)
 
 PY3 = sys.version_info >= (3, 0)
+PY35 = sys.version_info >= (3, 5)
 
 if PY3:
     from functools import lru_cache  # noqa F401
@@ -26,6 +29,7 @@ else:
     unichar = unichr  # noqa
     string = basestring  # noqa
 
+_QUIRKS = 0x20000
 DEBUG = 0x10000
 
 LC_A = ord('a')
@@ -153,4 +157,40 @@ def warn_deprecated(message, stacklevel=2):  # pragma: no cover
         message,
         category=DeprecationWarning,
         stacklevel=stacklevel
+    )
+
+
+class QuirksWarning(UserWarning):  # pragma: no cover
+    """Warning for quirks mode."""
+
+
+def warn_quirks(message, recommend, pattern):
+    """Warn quirks."""
+
+    import traceback
+    import bs4  # noqa: F401
+
+    paths = (MODULE, sys.modules['bs4'].__path__[0])
+    tb = traceback.extract_stack()
+    previous = None
+    filename = None
+    lineno = None
+    for entry in tb:
+        if (PY35 and entry.filename.startswith(paths)) or (not PY35 and entry[0].startswith(paths)):
+            break
+        previous = entry
+    if previous:
+        filename = previous.filename if PY35 else previous[0]
+        lineno = previous.lineno if PY35 else previous[1]
+
+    warnings.warn_explicit(
+        "\nCSS selector pattern: {!r}\n".format(pattern) +
+        "    {}\n".format(message) +
+        "    This behavior is only allowed temporarily for Beautiful Soup's transition to Soup Sieve.\n" +
+        "    In order to confrom to the CSS spec, {}\n".format(recommend) +
+        "    It is strongly recommended the selector be altered to conform to the CSS spec " +
+        "as an exception will be raised for this case in the future.\n",
+        QuirksWarning,
+        filename,
+        lineno
     )