+++ tests/util.py
@@ -14,8 +14,53 @@ XHTML = 4
 XML = 8
 
 
+def skip_quirks(func):
+    """Decorator that skips when quirks mode is enabled."""
+
+    def skip_if(self, *args, **kwargs):
+        """Skip conditional wrapper."""
+        if self.quirks is True:
+            return
+        else:
+            return func(self, *args, **kwargs)
+    return skip_if
+
+
+def skip_no_quirks(func):
+    """Decorator that skips when no quirks mode is enabled."""
+
+    def skip_if(self, *args, **kwargs):
+        """Skip conditional wrapper."""
+        if self.quirks is False:
+            return
+        else:
+            return func(self, *args, **kwargs)
+    return skip_if
+
+
 class TestCase(unittest.TestCase):
     """Test case."""
+
+    def setUp(self):
+        """Setup."""
+
+        sv.purge()
+        self.quirks = False
+
+    def purge(self):
+        """Purge cache."""
+
+        sv.purge()
+
+    def assert_raises(self, pattern, exception, namespace=None):
+        """Assert raises."""
+
+        with self.assertRaises(exception):
+            flags = sv.DEBUG
+            if self.quirks:
+                flags = sv._QUIRKS
+
+            sv.compile(pattern, flags=flags)
 
     def assert_selector(self, markup, selectors, expected_ids, namespaces={}, flags=0):
         """Assert selector."""
@@ -29,8 +74,12 @@ class TestCase(unittest.TestCase):
             bs_mode = 'xml'
         soup = bs4.BeautifulSoup(textwrap.dedent(markup.replace('\r\n', '\n')), bs_mode)
 
+        flags = sv.DEBUG
+        if self.quirks:
+            flags = sv._QUIRKS
+
         ids = []
-        for el in sv.select(selectors, soup, namespaces=namespaces, flags=sv.DEBUG):
+        for el in sv.select(selectors, soup, namespaces=namespaces, flags=flags):
             print('TAG: ', el.name)
             ids.append(el.attrs['id'])
         self.assertEqual(sorted(ids), sorted(expected_ids))